# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/branches": {
            "get": {
                "tags": [
                    "Branches"
                ],
                "summary": "Consultar sucursales",
                "description": "Retorna todas las sucursales",
                "operationId": "76938a0089ac92cfb2e0fb40af32c73b",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Branches"
                ],
                "summary": "Crear sucursal",
                "description": "Crea una nueva sucursal",
                "operationId": "582a679c4adf05c5fdb745be9100e5d6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "direccion"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Sucursal Centro"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Av. Principal 123"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "555-1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Sucursal creada"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/branches/{branch}": {
            "get": {
                "tags": [
                    "Branches"
                ],
                "summary": "Consultar una sucursal",
                "description": "Retorna una sucursal específica",
                "operationId": "5ce397d53def55d3db5ea6c4d928f87d",
                "parameters": [
                    {
                        "name": "branch",
                        "in": "path",
                        "description": "ID de la sucursal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "404": {
                        "description": "Sucursal no encontrada"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Branches"
                ],
                "summary": "Actualizar sucursal",
                "description": "Actualiza una sucursal existente",
                "operationId": "111cd4c99318763a36068408c93ad026",
                "parameters": [
                    {
                        "name": "branch",
                        "in": "path",
                        "description": "ID de la sucursal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Sucursal Norte"
                                    },
                                    "direccion": {
                                        "type": "string",
                                        "example": "Calle 456"
                                    },
                                    "telefono": {
                                        "type": "string",
                                        "example": "555-5678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Sucursal actualizada"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Sucursal no encontrada"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Branches"
                ],
                "summary": "Eliminar sucursal",
                "description": "Elimina una sucursal existente",
                "operationId": "bb8d9aa1a9983fb94d8d1868d0da6ae4",
                "parameters": [
                    {
                        "name": "branch",
                        "in": "path",
                        "description": "ID de la sucursal",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Sucursal eliminada"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Sucursal no encontrada"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Consultar todas las órdenes",
                "description": "Retorna todas las órdenes",
                "operationId": "955bb77a45ae173a485fadb8befed93f",
                "responses": {
                    "200": {
                        "description": "Operación exitosa"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "No se encontraron órdenes"
                    },
                    "405": {
                        "description": "Método no permitido"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Crear orden",
                "description": "Crear una nueva orden",
                "operationId": "2123ba58dadb22bd6f2c8081af562230",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "branch_id",
                                    "fecha_pedido",
                                    "fecha_recogida",
                                    "estado",
                                    "total"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "branch_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "fecha_pedido": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-10-27"
                                    },
                                    "fecha_recogida": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-10-28"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "example": "pendiente"
                                    },
                                    "total": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150.5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Orden creada"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}